fix compiler warnings from "nitro" repo

+++ modules/c++/sys/unittests/test_os.cpp
@@ -169,7 +169,7 @@ TEST_CASE(testSplitEnv)
     std::vector<std::string> paths;
     bool result = os.splitEnv(pathEnvVar, paths);
     TEST_ASSERT_TRUE(result);
-    TEST_ASSERT_GREATER(paths.size(), 0);
+    TEST_ASSERT_GREATER(paths.size(), static_cast<size_t>(0));
     for (const auto& path : paths)
     {
         TEST_ASSERT_TRUE(fs::exists(path));
@@ -395,13 +395,13 @@ TEST_CASE(testFsFileSize)
     {
         const fs::path argv0(os.getSpecialEnv("ARGV0"));
         const auto size = file_size(argv0);
-        TEST_ASSERT_GREATER(size, 0);
+        TEST_ASSERT_GREATER(size, static_cast<size_t>(0));
     }
     {
         // We always have  sys::filesystem, even if it's not used.
         const sys::filesystem::path argv0(os.getSpecialEnv("ARGV0"));
         const auto size = file_size(argv0);
-        TEST_ASSERT_GREATER(size, 0);
+        TEST_ASSERT_GREATER(size, static_cast<size_t>(0));
     }
 }
